# [SofaSimpleFem][SofaGeneralSimpleFem][SofaMiscFem] Avoid vectors normalization