… examples/Components/forcefield/FastTetrahedralCorotationalForceField.scn