@alxbilger 2022-01-31
I benchmarked only TetrahedronFEMForceFieldScene:

Before
```
------------------------------------------------------------------------------------------------------------------------
Benchmark                                                              Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------------------------
BM_Scene_bench_StepFactor<TetrahedronFEMForceFieldScene>/512        3197 ms         3188 ms            1 FPS=160.627/s frame=6.22559ms
BM_Scene_bench_StepFactor<TetrahedronFEMForceFieldScene>/1024       6434 ms         6438 ms            1 FPS=159.068/s frame=6.28662ms
BM_Scene_bench_StepFactor<TetrahedronFEMForceFieldScene>/2048      12923 ms        12938 ms            1 FPS=158.3/s frame=6.31714ms
BM_Scene_bench_StepFactor<TetrahedronFEMForceFieldScene>/4096      26074 ms        26062 ms            1 FPS=157.161/s frame=6.36292ms
```

After

```
------------------------------------------------------------------------------------------------------------------------
Benchmark                                                              Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------------------------
BM_Scene_bench_StepFactor<TetrahedronFEMForceFieldScene>/512        3128 ms         3125 ms            1 FPS=163.84/s frame=6.10352ms
BM_Scene_bench_StepFactor<TetrahedronFEMForceFieldScene>/1024       6343 ms         6344 ms            1 FPS=161.419/s frame=6.19507ms
BM_Scene_bench_StepFactor<TetrahedronFEMForceFieldScene>/2048      12693 ms        12656 ms            1 FPS=161.817/s frame=6.17981ms
BM_Scene_bench_StepFactor<TetrahedronFEMForceFieldScene>/4096      25060 ms        25062 ms            1 FPS=163.431/s frame=6.11877ms
```

It's not very significative since it is run only once, but it tends to confirm that the changes do accelerate the simulations.