@alxbilger 2022-02-11
One other condition on the fact it is not necessary to normalize again after a cross product is that both input vectors must be orthogonal. This has to be checked before accepting this PR.